feat: update MessageDTO and Message classes to use int for IDs and add timestamp getter method

+++ src/main/java/org/kungnection/dto/MessageDTO.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class MessageDTO {
-    private Long id;
-    private Long senderId;
+    private int id;
+    private int senderId;
     private String senderName;
     private String content;
     private String timestamp; // 建議回傳 ISO 格式字串
